# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AiChatRequestLanguage = typeof AiChatRequestLanguage[keyof typeof AiChatRequestLanguage];


export const AiChatRequestLanguage = {
  ht: 'ht',
  fr: 'fr',
} as const;

export interface AiChatRequest {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  question: string;
  language: AiChatRequestLanguage;
  /** Optional base64 data URL for a future image-capable provider. */
  imageData?: string | null;
}

export interface AiChatResponse {
  message: string;
  provider: string;
  configured: boolean;
}

export interface ApiProblem {
  code: string;
  message: string;
}