# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ai
    description: Faroh IA assistant operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /ai/chat:
    post:
      operationId: sendAiChat
      tags: [ai]
      summary: Send a question to the Faroh IA assistant
      description: Sends a student question to the configured AI provider. Returns a clear configuration response while no provider key is available.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
        "400":
          description: Invalid question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "502":
          description: AI provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "503":
          description: AI provider is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AiChatRequest:
      type: object
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 4000
        language:
          type: string
          enum: [ht, fr]
        imageData:
          type: string
          nullable: true
          description: Optional base64 data URL for a future image-capable provider.
      required:
        - question
        - language
    AiChatResponse:
      type: object
      properties:
        message:
          type: string
        provider:
          type: string
        configured:
          type: boolean
      required:
        - message
        - provider
        - configured
    ApiProblem:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message