# lib/api-zod/src/generated/types/aiChatRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiChatRequestLanguage } from './aiChatRequestLanguage';

export interface AiChatRequest {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  question: string;
  language: AiChatRequestLanguage;
  /** Optional base64 data URL for a future image-capable provider. */
  imageData?: string | null;
}