# lib/api-zod/src/generated/types/aiChatResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiChatResponse {
  message: string;
  provider: string;
  configured: boolean;
}